profiles: move the @receivers to receivers.py
