+++ eahub/profiles/apps.py
@@ -4,3 +4,9 @@ from django import apps
 class ProfilesConfig(apps.AppConfig):
     name = "eahub.profiles"
     verbose_name = "Profiles"
+
+    # noinspection PyUnresolvedReferences
+    def ready(self):
+        from eahub.profiles.receivers import clear_the_cache
+        from eahub.profiles.receivers import on_change
+        from eahub.profiles.receivers import save_new_profile_to_analytics
